♻️ [승훈]refactor: styled-components 분리
